Extract common nav utils to prepare for Horizontal Nav
