+++ ui/apps/platform/src/Containers/MainPage/Navigation/NavigationContent.tsx
