+++ ui/apps/platform/src/Containers/MainPage/Navigation/NavigationItem.tsx
