+++ ui/apps/platform/src/Containers/MainPage/Navigation/NavigationSidebar.css
